move javalin annotations out of generator
