+++ http-api/src/main/java/module-info.java
@@ -1,6 +1,7 @@
 module io.avaje.http.api {
 
 	exports io.avaje.http.api;
+	exports io.avaje.http.api.javalin;
 	exports io.avaje.http.api.context;
 	exports io.avaje.http.api.spi;
 }
