+++ http-generator-javalin/pom.xml
@@ -23,6 +23,13 @@
       <version>${avaje.prisms.version}</version>
       <scope>provided</scope>
     </dependency>
+    <dependency>
+      <groupId>io.avaje</groupId>
+      <artifactId>avaje-http-api</artifactId>
+      <version>${project.version}</version>
+      <optional>true</optional>
+      <scope>provided</scope>
+    </dependency>
 
   </dependencies>
