+++ http-generator-javalin/src/main/java/io/avaje/http/generator/javalin/JavalinProcessor.java
@@ -2,12 +2,12 @@ package io.avaje.http.generator.javalin;
 
 import java.io.IOException;
 
+import io.avaje.http.api.javalin.After;
+import io.avaje.http.api.javalin.Before;
 import io.avaje.http.generator.core.BaseProcessor;
 import io.avaje.http.generator.core.ControllerReader;
 import io.avaje.http.generator.core.PlatformAdapter;
 import io.avaje.http.generator.core.ProcessingContext;
-import io.avaje.http.javalin.After;
-import io.avaje.http.javalin.Before;
 import io.avaje.prism.GeneratePrism;
 
 @GeneratePrism(value = After.class, superClass = AbstractCustomMethodPrism.class)
