+++ http-generator-javalin/src/main/java/module-info.java
@@ -6,6 +6,7 @@ module io.avaje.http.javalin.generator {
   requires java.sql;
 
   // SHADED: All content after this line will be removed at package time
-  requires transitive io.avaje.http.generator.core;
+  requires io.avaje.http.generator.core;
+  requires io.avaje.http.api;
   requires static io.avaje.prism;
 }
